Refactor booking page to suspense-wrapped client component
